chore: types
- add theme module override
- inject theme declarations in libs/app
- update sx props to use mui theme
- remove typescript strict
