+++ apps/ousd/src/main.tsx
@@ -1,18 +1,20 @@
 import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
-import { OUSDRoot } from './components';
-import { QueryClientProvider } from '@tanstack/react-query';
-import { queryClient } from '@origin/shared/data-access';
+
 import {
   CssBaseline,
   Experimental_CssVarsProvider as CssVarsProvider,
 } from '@mui/material';
+import { queryClient } from '@origin/shared/data-access';
 import { theme } from '@origin/shared/theme';
+import { QueryClientProvider } from '@tanstack/react-query';
 import { IntlProvider } from 'react-intl';
+
+import { OUSDRoot } from './components';
 import { en } from './lang';
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.getElementById('root') as HTMLElement,
 );
 root.render(
   <StrictMode>
@@ -24,5 +26,5 @@ root.render(
         </CssVarsProvider>
       </QueryClientProvider>
     </IntlProvider>
-  </StrictMode>
+  </StrictMode>,
 );
