+++ libs/oeth/swap/tsconfig.lib.json
@@ -6,7 +6,8 @@
   },
   "files": [
     "../../../node_modules/@nx/react/typings/cssmodule.d.ts",
-    "../../../node_modules/@nx/react/typings/image.d.ts"
+    "../../../node_modules/@nx/react/typings/image.d.ts",
+    "../../../libs/shared/theme/src/theme.d.ts"
   ],
   "exclude": [
     "jest.config.ts",
