+++ libs/shared/components/tsconfig.json
@@ -4,7 +4,6 @@
     "allowJs": false,
     "esModuleInterop": false,
     "allowSyntheticDefaultImports": true,
-    "strict": true,
     "types": ["vitest"]
   },
   "files": [],
