+++ libs/shared/storybook/tsconfig.json
@@ -3,14 +3,13 @@
   "compilerOptions": {
     "module": "commonjs",
     "forceConsistentCasingInFileNames": true,
-    "strict": true,
     "noImplicitOverride": true,
     "noPropertyAccessFromIndexSignature": true,
     "noImplicitReturns": true,
     "noFallthroughCasesInSwitch": true,
     "jsx": "react-jsx"
   },
-  "files": [],
+  "files": ["../../../libs/shared/theme/src/theme.d.ts"],
   "include": [],
   "references": [
     {
