+++ libs/shared/theme/src/theme.tsx
@@ -4,34 +4,6 @@ import {
 } from '@mui/material/styles';
 import shadows from '@mui/material/styles/shadows';
 
-declare module '@mui/material/styles' {
-  interface TypeBackground {
-    gradient1: string;
-    gradient2: string;
-    gradient3: string;
-    gradientSuccess: string;
-    gradientHover: string;
-    gradientHoverActionButton: string;
-  }
-
-  interface TypeBackgroundOptions {
-    gradient1: string;
-    gradient2: string;
-    gradient3: string;
-    gradientSuccess: string;
-    gradientHover: string;
-    gradientHoverActionButton: string;
-  }
-
-  interface Shape {
-    cardBorderRadius: number;
-  }
-
-  interface ShapeOptions {
-    cardBorderRadius: number;
-  }
-}
-
 export const theme = extendTheme({
   colorSchemes: {
     dark: {
