+++ libs/shared/theme/tsconfig.lib.json
@@ -6,7 +6,8 @@
   },
   "files": [
     "../../../node_modules/@nx/react/typings/cssmodule.d.ts",
-    "../../../node_modules/@nx/react/typings/image.d.ts"
+    "../../../node_modules/@nx/react/typings/image.d.ts",
+    "./src/theme.d.ts"
   ],
   "exclude": [
     "jest.config.ts",
